refactor(commands): extract 5 no-gate domains (reminders, model_perf, secrets, mcp, settings-consents)
Pure MOVEs, zero behavior change (the commands/pipeline.rs pattern):
commands/mod.rs -863 lines (38,073 -> 37,210), 5 new commands/*.rs, lib.rs
UNTOUCHED (all 41 moved #[tauri::command]s resolve via the pub use glob).

- reminders.rs: add_reminder(+_blocking), build_reminder_script,
  escape_applescript, parse_iso_ymd + their unit tests.
- model_perf.rs: the 21 on-device-model commands (brain/embed/ner model
  presence, posture, selection, RAM). The interleaved gated
  timeline/detail commands + the non-listed download_* commands were LEFT
  in mod.rs; each listed fn extracted individually.
- secrets.rs (#[path] alias — name shadows crate::secrets): the jira/
  slack/anthropic/gateway/web-search token setters+getters.
- mcp.rs (#[path] alias): the 5 MCP-server config commands (config rows +
  a fail-closed consented bool, no meeting content).
- settings.rs (#[path] alias): the 5 self-contained consent-flag commands
  (cloud-egress/web-search/jira/slack). The entangled config-DTO core
  (AppConfigDto + save/get_config + *_inner) was LEFT in mod.rs (a clean
  move needs relocating its non-listed _inner cores too — a follow-up).

No lock surface: verified per-fn that NO moved symbol touches
meeting_is_unlocked / visibility_clause / seal / folder_is_unlocked (the
consent flags are the dedicated flag-writers, not content reads; the
egress gate stays in the provider/connector layer). Tests stay in mod.rs;
#[test] parity 349+2 tokio on trunk and now. Verified: targeted cargo
test --lib green, clippy --lib -D warnings clean, MSRV grep clean.
